Fix handling of keyDown/keyUp events by TextInput
This extends the ability to intercept `keyDown` and `keyUp` events to `TextInput`. We need this for the ability to insert newlines when holding shift in chat, along with support arrow up/down from the search input.

This also fixes an issue where only one modifier flag was being looked at, so key events would not accurately show all modifier keys being held down but instead just have one of them.